feat: fix #31
reduced limit for methods per class
extracted defaultUserTests
