+++ config/pmd/ruleset.xml
@@ -31,6 +31,7 @@
         <exclude name="CommentDefaultAccessModifier"/>
         <exclude name="MethodNamingConventions"/>
         <exclude name="LinguisticNaming"/>
+        <exclude name="TooManyStaticImports"/>
     </rule>
     <rule ref="category/java/design.xml">
         <exclude name="LawOfDemeter"/>
@@ -48,7 +49,7 @@
     </rule>
     <rule ref="category/java/design.xml/TooManyMethods">
         <properties>
-            <property name="maxmethods" value="30"/>
+            <property name="maxmethods" value="25"/>
         </properties>
     </rule>
     <rule ref="category/java/design.xml/TooManyFields">
